Adds autologin support
A new configuration option allows users to configure auth_saml2 to
attempt passive login automatically, without the user having to go
to a login page.

This is useful for servers that allow users to access without being
logged in (e.g. public access), but where some users might be
already logged on to the institutional identity provider.

There are two options: try at the start of a session, or try every
time a cookie changes. The first option is simple but means it won't
detect login if you go to the server before logging in. The second
option can be useful if you are able to configure your identity
provider to set a cookie within the domain used by the Moodle
system when the user logs in.

In both cases, it will not do anything when you are already logged
in (this is an autologin feature, not autologout).